Create homework-1
Following the Javascript is sexy course load, This is Homework #1. A bit tough and did take some time to understand what was going on (with help), my biggest hump to overcome was saving the values to the array via the .push(); method and then doing something with them. Thank goodness for psuedo-code.